+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+1.2.0 (2019-06-11)
+------------------
+
+* Code now adds a citation attribute to every edge even if there is no value
+  in which case an empty list is set (JIRA ticket UD-360)
+
 1.1.0 (2019-06-10)
 ------------------